Immplement shortcut handling.
