+++ app/gui/src/presenter/project.rs
@@ -278,6 +278,10 @@ impl Model {
             view.show_graph_editor();
         })
     }
+
+    fn execution_mode_changed(&self, mode: &ide_view::execution_mode_selector::ExecutionMode) {
+        error!("Execution mode changed: {mode}");
+    }
 }
 
 
@@ -374,6 +378,8 @@ impl Project {
             eval_ view.execution_context_interrupt(model.execution_context_interrupt());
 
             eval_ view.execution_context_restart(model.execution_context_restart());
+
+            eval graph_view.execution_mode((mode) model.execution_mode_changed(mode));
         }
 
         let graph_controller = self.model.graph_controller.clone_ref();
@@ -388,7 +394,7 @@ impl Project {
     /// implementation of #5930.
     fn init_execution_modes(self) -> Self {
         let graph = &self.model.view.graph();
-        let entries = Rc::new(vec!["development".to_string(), "production".to_string()]);
+        let entries = Rc::new(vec!["design".to_string(), "live".to_string()]);
         graph.set_available_execution_modes(entries);
         self
     }
